use IOtest for python matrix element
